// node/src/osparc-api-ts-client/models/SolverFunctionJob.ts
/**
 * osparc.io public API
 * osparc-simcore public API specifications
 *
 * OpenAPI spec version: 0.8.0.post0.dev0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class SolverFunctionJob {
    'title'?: string;
    'description'?: string;
    'functionUid': string;
    'inputs': any | null;
    'outputs': any | null;
    'functionClass'?: SolverFunctionJobFunctionClassEnum;
    'solverJobId': string;
    'status': string; // manually added
    'uid': string; // manually added

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "functionUid",
            "baseName": "function_uid",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "any",
            "format": ""
        },
        {
            "name": "outputs",
            "baseName": "outputs",
            "type": "any",
            "format": ""
        },
        {
            "name": "functionClass",
            "baseName": "function_class",
            "type": "SolverFunctionJobFunctionClassEnum",
            "format": ""
        },
        {
            "name": "solverJobId",
            "baseName": "solver_job_id",
            "type": "string",
            "format": "uuid"
        }    ];

    static getAttributeTypeMap() {
        return SolverFunctionJob.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum SolverFunctionJobFunctionClassEnum {
    Solver = 'SOLVER'
}